requirements.txt
selenium==4.15.2
webdriver-manager==4.0.0
beautifulsoup4==4.12.2
pandas==2.1.0
numpy==1.24.3
faiss-cpu==1.7.4
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
google-generativeai==0.3.0
streamlit==1.27.0
python-multipart==0.0.6
requests==2.31.0
